Fixing PR issues and removing redundant model
