--- app/src/main/java/br/com/concrete/tentacle/data/models/RegisteredMediaResponse.kt
@@ -1,10 +1,0 @@
-package br.com.concrete.tentacle.data.models
-
-data class RegisteredMediaResponse(
-    val _id: String,
-    val platform: String,
-    val game: Game,
-    val owner: User,
-    val createdAt: String,
-    val updateAt: String
-)
