+++ app/src/main/java/br/com/concrete/tentacle/features/registerGame/registerMedia/RegisterMediaViewModel.kt
@@ -5,15 +5,15 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.OnLifecycleEvent
 import br.com.concrete.tentacle.base.BaseViewModel
 import br.com.concrete.tentacle.data.models.Game
+import br.com.concrete.tentacle.data.models.Media
 import br.com.concrete.tentacle.data.models.MediaRequest
-import br.com.concrete.tentacle.data.models.RegisteredMediaResponse
 import br.com.concrete.tentacle.data.models.ViewStateModel
 import br.com.concrete.tentacle.data.repositories.RegisterMediaRepository
 import br.com.concrete.tentacle.utils.LogWrapper
 
 class RegisterMediaViewModel(private val repository: RegisterMediaRepository) : BaseViewModel() {
 
-    val viewStatusModel = MutableLiveData<ViewStateModel<RegisteredMediaResponse>>()
+    val viewStatusModel = MutableLiveData<ViewStateModel<Media>>()
 
     fun registerMedia(platform: String, game: Game) {
         viewStatusModel.postValue(ViewStateModel(ViewStateModel.Status.LOADING))
